rename gmock-qofbook.h and gmock-Split.h to .hpp
