+++ libgnucash/engine/mocks/gmock-qofbook.hpp
